requirements.txt

dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-environ==0.4.5
django-filter==2.3.0
django-graphql-auth==0.3.11
django-graphql-jwt==0.3.0
django-heroku==0.3.1
django-sequences==2.5
django-storages==1.9.1
django-userforeignkey==0.3.0
djangorestframework==3.11.0
fcm-django==0.3.4
gitdb==4.0.7
GitPython==3.1.14
mysqlclient==2.0.1
PyGithub==1.54.1
PyMySQL==0.10.0





